Added support for developerPayload on android
DeveloperPayload is optionally used with subscriptions and
consumables/non consumables on Android. iOS does not support payloads.

Warning: It’s not validated in any way inside of the plugin. It’s
created only to be able to pass payload as part of order and to be able
to validate it on server.